refactoring and adding standfirst to tests
