+++ index.js
@@ -4,7 +4,7 @@ const NewsView = require('./src/newsView');
 
 const client = new NewsClient();
 const model = new NewsModel();
-const news = require('./src/mockNews.js');
+const news = require('./mock/mockNews.js');
 model.setNews(news);
 
 const view = new NewsView(model, client);
