+++ src/newsClient.js
@@ -3,7 +3,7 @@ const API_KEY = require('./apiKey.js');
 class NewsClient {
   loadNews(searchTerm = 'uk', callback) {
     fetch(
-      `https://content.guardianapis.com/search?q=${searchTerm}&query-fields=headline&show-fields=thumbnail,headline,byline&order-by=newest&api-key=${API_KEY}`
+      `https://content.guardianapis.com/search?q=${searchTerm}&query-fields=headline&show-fields=thumbnail,headline,byline,standfirst&order-by=newest&api-key=${API_KEY}`
     )
       .then((response) => response.json())
       .then((data) => {
